Improvements in getattr and statfs: du will not return meaningful values and df will return something which is not as wrong as it was (but, still not correct).
Fuse::DBI will not try to load kernel module using sudo, and try to umount
using sudo umount if fusermount -u fails (as it happends with current CVS
version of fuse).

New webgui test target in Makefile which work as test on my local machine
(and hopefully on any with webgui default installation).


git-svn-id: svn://svn.rot13.org/fuse_dbi/trunk@51 17f4e80c-d0e0-0310-8903-bfc3ae804c12